VenueManagerDAO added and implemented update venue in RegisterVenueDAO
