+++ src/javaapplication6/dao/UserDAO.java
@@ -182,22 +182,4 @@ public class UserDAO {
            return false;
         }
     }
-    public boolean updateVenueDetails(VenueModel venue) {
-        String sqlQuery = "UPDATE venue_table SET name = ?, location = ?, email = ?, contact_number = ?, price_per_plate = ?, status = ? WHERE id = ?";
-        try (Connection conn = dbConn.connection_base()) {
-            PreparedStatement pstmt = conn.prepareStatement(sqlQuery);
-            pstmt.setString(1, venue.getName());
-            pstmt.setString(2, venue.getLocation());
-            pstmt.setString(3, venue.getEmail());
-            pstmt.setString(4, venue.getContact_number());
-            pstmt.setDouble(5, venue.getPrice_per_plate());
-            pstmt.setString(6, venue.getStatus());
-            pstmt.setInt(7, venue.getId());
-
-            return pstmt.executeUpdate() > 0;
-        } catch (Exception e) {
-            e.printStackTrace();
-            return false;
-        }
-    }
 }
